Demote LAMP applications
With the docs for the two mentioned LAMP applications now demoted, there
is little reason to keep mention of LAMP in general.  This is an old web
application architectural strategy that, while still viable, is not
considered a best practice, and not what we would promote for new web
developers who use Ubuntu.

These days, there are a myriad number of other web application
systems, with new ones evolving into and out of existance over a period
of just a few years.  As a result, Ubuntu Server has refrained from
picking one stack or framework to push.  Instead, we focus on the
individual underlying components (web servers, databases, etc.) that are
common to many stacks and frameworks.