ssh: fix service reference
The old used to work and will work again after [1] is fixed,
but the ssh service on disk is /usr/lib/systemd/system/ssh.service
and sshd only an alias (which currently happens to be broken on noble).
Using ssh.service in our examples avoids readers to stumble until
fixed in an SRU and can stay as-is afterwards.

[1]: https://bugs.launchpad.net/ubuntu/+source/openssh/+bug/2087949

Fixes: #93

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>